Splitting DepositHandler into multiple files
